Address CoPilot comments
